requirements.txt
pillow~=10.4.0
pip~=22.0.2
setuptools~=59.6.0